requirements.txt
fastapi>=0.104.0
pyyaml>=6.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.4.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-cov>=4.1.0
